refactor: Completely separated buisness logics from UI
